requirements.txt
# Core environment
numpy
scikit-learn
opencv-python
pillow
torch
torchvision

# Face recognition & detection
facenet-pytorch
ultralytics     

# Util
